installer/output/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
bleach==6.1.0
celery==5.3.4
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==2024.10
cron_descriptor==2.0.6
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.7
django-axes==6.1.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-crispy-forms==2.3
django-defender==0.9.7
django-filter==23.5
django-import-export==3.3.5
django-ratelimit==4.1.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
django_csp==3.8
djangorestframework==3.14.0
et_xmlfile==2.0.0
gunicorn==21.2.0
kombu==5.5.4
MarkupPy==1.18
mysqlclient==2.2.7
odfpy==1.4.1
openpyxl==3.1.2
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycparser==2.23
python-crontab==3.3.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
reportlab==4.0.7
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tablib==3.5.0
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.6.0
xlrd==2.0.2
xlwt==1.3.0
requests==2.31.0
